added metalabel gen

+++ PhyTrade/Data_Collection_preparation/Record_splines.py
@@ -1,5 +1,7 @@
 from PhyTrade.Tools.INDIVIDUAL_gen import Individual
 import pandas as pd
+import json
+import os.path
 
 
 def record_splines(parameter_set, data_slice, ticker, spline_type=None):
@@ -17,22 +19,24 @@ def record_splines(parameter_set, data_slice, ticker, spline_type=None):
     individual = Individual(ticker=ticker, parameter_set=parameter_set)
     individual.gen_economic_model(data_slice)
 
-    index = data_slice.data[data_slice.default_start_index:data_slice.default_end_index]["Date"]
-    spline_df = pd.DataFrame(index=index, columns=["trade_spline", "trade_signal"])
+    spline_df = pd.DataFrame(columns=["Date", "trade_spline", "trade_signal"])
 
-    # TODO: Fixed length index error
-
+    spline_df["Date"] = data_slice.data[data_slice.start_index:data_slice.stop_index]["Date"]
     spline_df["trade_spline"] = individual.trade_spline
     spline_df["trade_signal"] = individual.trade_signal
 
     # ---> Save spline to csv file
     if spline_type is None:
-        path = r"C:\Users\Victor Guillet\Google Drive\2-Programing\Repos\Python\Steffegium\Data\Splines".replace(
-            '\\', '/')
+        path = r"C:\Users\Victor Guillet\Google Drive\2-Programing\Repos\Python\Steffegium\Data\Splines".replace('\\', '/')
     else:
         path = r"C:\Users\Victor Guillet\Google Drive\2-Programing\Repos\Python\Steffegium\Data\Metalabeling_results\**".\
             replace('\\', '/').replace('**', spline_type)
 
-    full_file_name = path + '/' + ticker + "_splines_" + data_slice.default_start_date + " - " + data_slice.default_end_date
-    spline_df.to_csv(full_file_name)
+    full_file_name = path + '/' + ticker + "_splines"
 
+    if os.path.isfile(full_file_name):
+        new_df = pd.concat([pd.read_csv(full_file_name, index_col=0), spline_df])
+        new_df.to_csv(full_file_name)
+
+    else:
+        spline_df.to_csv(full_file_name)
